+++ packages/vite/src/index.ts
@@ -1,10 +1,15 @@
 
-import { supercharge } from './plugin.js'
+import { supercharge } from './plugin/plugin.js'
 
 export default supercharge
 export { supercharge }
 
-export { PluginConfigContract, DevServerUrl } from './contracts/index.js'
-export { resolvePageComponent } from './inertia-helpers.js'
-export { Vite, Attributes } from './vite.js'
+export { Vite, ViteTagAttributes } from './backend/vite.js'
+export { ViteConfig } from './backend/vite-config.js'
 export { ViteServiceProvider } from './vite-service-provider.js'
+
+export { resolvePageComponent } from './inertia/inertia-helpers.js'
+export { InertiaPageNotFoundError } from './inertia/inertia-page-not-found-error.js'
+
+export { HotReloadFile } from './plugin/hotfile.js'
+export { PluginConfigContract, DevServerUrl } from './plugin/types.js'