+++ packages/vite/src/inertia/inertia-helpers.ts
@@ -1,4 +1,6 @@
 
+import { InertiaPageNotFoundError } from './inertia-page-not-found-error.js'
+
 /**
  * Resolves the inertia page component for the given `path` from the available `pages`.
  *
@@ -16,10 +18,10 @@ export async function resolvePageComponent<T> (path: string, pages: Record<strin
   const page = pages[path]
 
   if (page == null) {
-    throw new Error(`Inertia page not found: ${path}`)
+    throw new InertiaPageNotFoundError(path)
   }
 
   return typeof page === 'function'
-    ? page()
+    ? await page()
     : page
 }